Extract live subscription publisher
